Moved all files into src package for better organization
