+++ alembic/env.py
@@ -6,8 +6,8 @@ from sqlalchemy import pool
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
-from config import DATABASE_URI
-from models import BaseModel
+from src.config import DATABASE_URI
+from src.models import BaseModel
 
 config = context.config
