--- bot/__init__.py
@@ -1,3 +1,0 @@
-from containers import WIRE_TO
-
-WIRE_TO.append("bot.controllers")
