+++ src/bot/__init__.py
@@ -1,0 +1,3 @@
+from src.containers import WIRE_TO
+
+WIRE_TO.append("src.bot.controllers")
